refactor(test): extract transition helpers to lib to fix Next.js async Server Action constraint
